Cleaned code up

+++ src/app/app.component.ts
@@ -5,18 +5,16 @@ import {SplashScreen} from '@ionic-native/splash-screen';
 
 import {AuthPage} from "../pages/auth/auth";
 import {AuthService} from "../services/auth.service";
-import {SettingsPage} from "../pages/settings/settings";
 import {User} from "../models/user.model";
 import {TabsPage} from "../pages/tabs/tabs";
 import {StorageService} from "../services/storage.service";
-import {SigninModal} from "../pages/modals/signin-modal/signin-modal";
+import {SigninModal} from "../pages/modals/signin-modal/signin.modal";
 import {HelperService} from "../services/helpers";
 
 export {TabsPage} from '../pages/tabs/tabs';
 export {GraphqlPage} from '../pages/graphql/graphql';
 export {RestfulPage} from '../pages/restful/restful';
 export {AuthPage} from '../pages/auth/auth';
-export {SettingsPage} from '../pages/settings/settings';
 export {HistoryPage} from '../pages/history/history';
 export {GraphPage} from '../pages/history/graph/graph';
 export {RestPage} from '../pages/history/rest/rest';
@@ -26,7 +24,7 @@ export {RequestHeaderComponent} from '../components/header';
 export {RequestBodyComponent} from '../components/body';
 export {PopoverComponent} from '../components/popover';
 export {RequestAuthorizationComponent} from '../components/authorization';
-export {SigninModal} from '../pages/modals/signin-modal/signin-modal';
+export {SigninModal} from '../pages/modals/signin-modal/signin.modal';
 
 @Component({
     templateUrl: 'app.html'
@@ -34,7 +32,7 @@ export {SigninModal} from '../pages/modals/signin-modal/signin-modal';
 export class Entry implements OnInit {
     private _date = new Date();
     authPage = AuthPage;
-    settingsPage = SettingsPage;
+    signinModal = SigninModal;
     user: User;
     @ViewChild('nav') nav: NavController;
 
@@ -47,6 +45,7 @@ export class Entry implements OnInit {
      * @param authService
      * @param storageService
      * @param h
+     * @param logger
      */
     constructor(private _platform: Platform,
                 private _statusBar: StatusBar,
@@ -59,6 +58,10 @@ export class Entry implements OnInit {
         this._initializeApp();
     }
 
+    /**
+     * Initialize Altfire Application
+     * @private
+     */
     _initializeApp() {
         this._platform.ready().then(() => {
             this._statusBar.styleDefault();
@@ -73,24 +76,28 @@ export class Entry implements OnInit {
     ngOnInit() {
         let currentTime = this._date.getTime();
         this.storageService.renderData();
-        this.storageService.getAltfireApp().then((data) => {
+        this.storageService.getAltfireApp().then(() => {
             this.storageService.getUser().then((user: User) => {
-                if(user.token && user.expire > currentTime) {
+                if(user.uid && user.token && user.expire > currentTime) {
+                    this.showToast('Welcome back ' + user.username + ', missed you!');
                     user.tokenValid = true;
+                    // TODO: Check if there's a way to update data in LocalStorage for Ionic
                     this.authService.setUser(user);
                     this.nav.setRoot(TabsPage);
-                } else if (user.token && user.expire <= currentTime) {
+                } else if (user.uid && user.token && user.expire <= currentTime) {
+                    // TODO: Is the logic below necessary??
                     this.h.loader({msg: 'signing you in . . .'});
-                    let modal = this.modalCtrl.create(SigninModal, user);
-                    modal.onDidDismiss((data) => {
-                        if(data && data == 'continue') {
+                    let modal = this.modalCtrl.create(this.signinModal, user);
+                    modal.onDidDismiss((signinAs) => {
+                        if(signinAs && signinAs == 'olduser') {
                             user.tokenValid = true;
-                            this.onShowToast('Logged in as ' + user.username);
+                            this.showToast('Logged in as ' + user.username);
                             this.authService.setUser(user);
+                            this.nav.setRoot(TabsPage);
                         } else {
                             this.storageService.deleteToken().then(() => {
                                 user.tokenValid = false;
-                                this.onShowToast('please signin to save your work');
+                                this.showToast('please signin to save your work');
                                 this.nav.setRoot(this.authPage, {auth_type: 'signin', data: this._initInput(user)});
                             }).catch((err) => {
                                 this.logger.handleError(err);
@@ -125,7 +132,7 @@ export class Entry implements OnInit {
      * @param data
      * @param duration
      */
-    onShowToast(data: string, duration?: number) {
+    showToast(data: string, duration?: number) {
         this.h.toast({msg: data, duration: duration ? duration : 2000}).present();
     }
